requirements.txt
libscrc
pyserial
